morphantic_engine_server/requirements.txt
fastapi
uvicorn[standard]
requests
numpy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
pymongo>=4.6.0
# Add any other specific dependencies your Morphantic engine needs
# Core Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning & Deep Learning
scikit-learn>=1.0.0
optuna>=3.0.0
yfinance>=0.2.0

# Chemistry & Drug Discovery
rdkit>=2022.03.1
DeepChem>=2.7.0

# Graph & Network Analysis

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Data Processing
tqdm>=4.62.0
joblib>=1.1.0

# Configuration & Utilities
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Optimization

# Testing (optional)
pytest>=7.0.0
pytest-cov>=3.0.0
